requirements.txt
streamlit
numpy
pandas
<<<<<<< HEAD
pymupdf
python-dotenv
=======
>>>>>>> 12df44253ebbda89ead07a4fa6652b545a6a9ebf
